+++ indico/core/signals/event/registration.py
@@ -96,3 +96,11 @@ Called when composing lists of badge templates. The `sender` may be either
 The list of badge templates is passed in the `badge_templates` kwarg.
 The signal handler is expected to mutate the list.
 ''')
+
+before_check_registration_email = _signals.signal('before-check-registration-email', '''
+Called before checking the validity of the registration email.  The sender is
+the ``RegistrationForm`` object.  The signal handler is expected to return
+``None`` if all checks passed or a ``{'status': ..., 'conflict': ...}``
+dictionary.  ``'status'`` is expected to be either ``'error'``, ``'warning'`` or
+``ok``.
+''')